+++ fixtures/Compactor/DummyCompactor.php
@@ -14,8 +14,6 @@ declare(strict_types=1);
 
 namespace KevinGH\Box\Compactor;
 
-use KevinGH\Box\Compactor;
-
 class DummyCompactor implements Compactor
 {
     /**
