+++ fixtures/Compactor/FakeCompactor.php
@@ -14,9 +14,8 @@ declare(strict_types=1);
 
 namespace KevinGH\Box\Compactor;
 
-use function func_get_args;
-use KevinGH\Box\Compactor;
 use KevinGH\Box\NotCallable;
+use function func_get_args;
 
 class FakeCompactor implements Compactor
 {
