+++ src/Compactor/BaseCompactor.php
@@ -14,8 +14,6 @@ declare(strict_types=1);
 
 namespace KevinGH\Box\Compactor;
 
-use KevinGH\Box\Compactor;
-
 /**
  * Base compactor class providing a slightly simpler API to compact the content only if the file is supported.
  *
