+++ src/Compactor/Placeholder.php
@@ -14,9 +14,8 @@ declare(strict_types=1);
 
 namespace KevinGH\Box\Compactor;
 
-use function array_keys;
 use Assert\Assertion;
-use KevinGH\Box\Compactor;
+use function array_keys;
 use function str_replace;
 
 final class Placeholder implements Compactor
