+++ src/Configuration/Configuration.php
@@ -12,8 +12,43 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace KevinGH\Box;
+namespace KevinGH\Box\Configuration;
 
+use Assert\Assertion;
+use Closure;
+use DateTimeImmutable;
+use DateTimeZone;
+use Herrera\Box\Compactor\Json as LegacyJson;
+use Herrera\Box\Compactor\Php as LegacyPhp;
+use Humbug\PhpScoper\Configuration as PhpScoperConfiguration;
+use Humbug\PhpScoper\Console\ApplicationFactory;
+use Humbug\PhpScoper\Scoper;
+use Humbug\PhpScoper\Scoper\FileWhitelistScoper;
+use InvalidArgumentException;
+use KevinGH\Box\Annotation\AnnotationDumper;
+use KevinGH\Box\Annotation\DocblockAnnotationParser;
+use KevinGH\Box\Annotation\DocblockParser;
+use KevinGH\Box\Compactor\Compactor;
+use KevinGH\Box\Compactor\Json as JsonCompactor;
+use KevinGH\Box\Compactor\Php as PhpCompactor;
+use KevinGH\Box\Compactor\PhpScoper as PhpScoperCompactor;
+use KevinGH\Box\Composer\ComposerConfiguration;
+use KevinGH\Box\Composer\ComposerFile;
+use KevinGH\Box\Composer\ComposerFiles;
+use KevinGH\Box\Json\Json;
+use KevinGH\Box\MapFile;
+use KevinGH\Box\PhpScoper\SimpleScoper;
+use KevinGH\Box\scalar;
+use Phar;
+use RuntimeException;
+use Seld\JsonLint\ParsingException;
+use SplFileInfo;
+use stdClass;
+use Symfony\Component\Finder\Finder;
+use Symfony\Component\Finder\SplFileInfo as SymfonySplFileInfo;
+use Symfony\Component\Process\Process;
+use Symfony\Component\VarDumper\Cloner\VarCloner;
+use Symfony\Component\VarDumper\Dumper\CliDumper;
 use function array_diff;
 use function array_filter;
 use function array_flip;
@@ -24,29 +59,16 @@ use function array_merge;
 use function array_unique;
 use function array_values;
 use function array_walk;
-use Assert\Assertion;
-use Closure;
 use function constant;
 use function current;
-use DateTimeImmutable;
-use DateTimeZone;
 use function defined;
 use function dirname;
-use const E_USER_DEPRECATED;
 use function explode;
-use File;
 use function file_exists;
 use function getcwd;
-use Herrera\Box\Compactor\Json as LegacyJson;
-use Herrera\Box\Compactor\Php as LegacyPhp;
-use Humbug\PhpScoper\Configuration as PhpScoperConfiguration;
-use Humbug\PhpScoper\Console\ApplicationFactory;
-use Humbug\PhpScoper\Scoper;
-use Humbug\PhpScoper\Scoper\FileWhitelistScoper;
 use function implode;
 use function in_array;
 use function intval;
-use InvalidArgumentException;
 use function is_array;
 use function is_bool;
 use function is_file;
@@ -57,45 +79,25 @@ use function is_string;
 use function iter\map;
 use function iter\toArray;
 use function iter\values;
-use KevinGH\Box\Annotation\AnnotationDumper;
-use KevinGH\Box\Annotation\DocblockAnnotationParser;
-use KevinGH\Box\Annotation\DocblockParser;
-use KevinGH\Box\Compactor\Json as JsonCompactor;
-use KevinGH\Box\Compactor\Php as PhpCompactor;
-use KevinGH\Box\Compactor\PhpScoper as PhpScoperCompactor;
-use KevinGH\Box\Composer\ComposerConfiguration;
-use KevinGH\Box\Composer\ComposerFile;
-use KevinGH\Box\Composer\ComposerFiles;
 use function KevinGH\Box\FileSystem\canonicalize;
 use function KevinGH\Box\FileSystem\file_contents;
 use function KevinGH\Box\FileSystem\is_absolute_path;
 use function KevinGH\Box\FileSystem\longest_common_base_path;
 use function KevinGH\Box\FileSystem\make_path_absolute;
 use function KevinGH\Box\FileSystem\make_path_relative;
-use KevinGH\Box\Json\Json;
-use KevinGH\Box\PhpScoper\SimpleScoper;
 use function krsort;
-use Phar;
 use function preg_match;
 use function preg_replace;
 use function property_exists;
 use function realpath;
-use RuntimeException;
-use Seld\JsonLint\ParsingException;
 use function sort;
-use const SORT_STRING;
-use SplFileInfo;
 use function sprintf;
-use stdClass;
 use function strtoupper;
 use function substr;
-use Symfony\Component\Finder\Finder;
-use Symfony\Component\Finder\SplFileInfo as SymfonySplFileInfo;
-use Symfony\Component\Process\Process;
-use Symfony\Component\VarDumper\Cloner\VarCloner;
-use Symfony\Component\VarDumper\Dumper\CliDumper;
 use function trigger_error;
 use function trim;
+use const E_USER_DEPRECATED;
+use const SORT_STRING;
 
 /**
  * @private
