+++ src/Console/Command/Compile.php
@@ -15,14 +15,34 @@ declare(strict_types=1);
 namespace KevinGH\Box\Console\Command;
 
 use Amp\MultiReasonException;
+use Assert\Assertion;
+use DateTimeImmutable;
+use DateTimeZone;
+use KevinGH\Box\Box;
+use KevinGH\Box\Compactor\Compactor;
+use KevinGH\Box\Composer\ComposerConfiguration;
+use KevinGH\Box\Configuration\Configuration;
+use KevinGH\Box\Console\Logger\CompileLogger;
+use KevinGH\Box\Console\MessageRenderer;
+use KevinGH\Box\Console\OutputFormatterConfigurator;
+use KevinGH\Box\Console\Php\PhpSettingsHandler;
+use KevinGH\Box\MapFile;
+use KevinGH\Box\RequirementChecker\RequirementsDumper;
+use KevinGH\Box\StubGenerator;
+use RuntimeException;
+use stdClass;
+use Symfony\Component\Console\Helper\QuestionHelper;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Input\StringInput;
+use Symfony\Component\Console\Logger\ConsoleLogger;
+use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\Question;
+use Symfony\Component\Console\Style\SymfonyStyle;
 use function array_map;
 use function array_search;
 use function array_shift;
-use Assert\Assertion;
 use function count;
-use const DATE_ATOM;
-use DateTimeImmutable;
-use DateTimeZone;
 use function decoct;
 use function explode;
 use function file_exists;
@@ -32,15 +52,6 @@ use function get_class;
 use function get_loaded_extensions;
 use function implode;
 use function is_string;
-use KevinGH\Box\Box;
-use const KevinGH\Box\BOX_ALLOW_XDEBUG;
-use KevinGH\Box\Compactor;
-use KevinGH\Box\Composer\ComposerConfiguration;
-use KevinGH\Box\Configuration;
-use KevinGH\Box\Console\Logger\CompileLogger;
-use KevinGH\Box\Console\MessageRenderer;
-use KevinGH\Box\Console\OutputFormatterConfigurator;
-use KevinGH\Box\Console\Php\PhpSettingsHandler;
 use function KevinGH\Box\disable_parallel_processing;
 use function KevinGH\Box\FileSystem\chmod;
 use function KevinGH\Box\FileSystem\dump_file;
@@ -50,34 +61,23 @@ use function KevinGH\Box\FileSystem\rename;
 use function KevinGH\Box\format_size;
 use function KevinGH\Box\get_box_version;
 use function KevinGH\Box\get_phar_compression_algorithms;
-use KevinGH\Box\MapFile;
-use KevinGH\Box\RequirementChecker\RequirementsDumper;
-use KevinGH\Box\StubGenerator;
 use function memory_get_peak_usage;
 use function memory_get_usage;
 use function microtime;
-use const PHP_EOL;
-use const PHP_VERSION;
 use function posix_getrlimit;
-use const POSIX_RLIMIT_INFINITY;
-use const POSIX_RLIMIT_NOFILE;
 use function posix_setrlimit;
 use function putenv;
 use function round;
-use RuntimeException;
 use function sprintf;
-use stdClass;
 use function strlen;
 use function substr;
-use Symfony\Component\Console\Helper\QuestionHelper;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Input\StringInput;
-use Symfony\Component\Console\Logger\ConsoleLogger;
-use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Question\Question;
-use Symfony\Component\Console\Style\SymfonyStyle;
 use function var_export;
+use const DATE_ATOM;
+use const KevinGH\Box\BOX_ALLOW_XDEBUG;
+use const PHP_EOL;
+use const PHP_VERSION;
+use const POSIX_RLIMIT_INFINITY;
+use const POSIX_RLIMIT_NOFILE;
 
 /**
  * @final
