+++ src/Verifier/Verifier.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace KevinGH\Box;
+namespace KevinGH\Box\Verifier;
 
 /**
  * Defines how a signature verification class must be implemented.
