+++ tests/BoxTest.php
@@ -14,44 +14,45 @@ declare(strict_types=1);
 
 namespace KevinGH\Box;
 
-use function array_filter;
-use function array_keys;
-use function current;
-use const DIRECTORY_SEPARATOR;
-use function dirname;
 use Exception;
-use function exec;
-use function extension_loaded;
-use function file_get_contents;
 use Generator;
-use function implode;
-use function in_array;
 use InvalidArgumentException;
-use function iterator_to_array;
+use KevinGH\Box\Compactor\Compactor;
 use KevinGH\Box\Compactor\FakeCompactor;
 use KevinGH\Box\Console\DisplayNormalizer;
-use function KevinGH\Box\FileSystem\canonicalize;
-use function KevinGH\Box\FileSystem\chmod;
-use function KevinGH\Box\FileSystem\dump_file;
-use function KevinGH\Box\FileSystem\make_tmp_dir;
-use function KevinGH\Box\FileSystem\mkdir;
 use KevinGH\Box\Test\FileSystemTestCase;
 use KevinGH\Box\Test\RequiresPharReadonlyOff;
 use org\bovigo\vfs\vfsStream;
 use org\bovigo\vfs\vfsStreamWrapper;
 use Phar;
 use PharFileInfo;
-use const PHP_EOL;
 use Prophecy\Argument;
 use Prophecy\Prophecy\ObjectProphecy;
-use function realpath;
 use RuntimeException;
 use SplFileInfo;
-use function sprintf;
-use const STDOUT;
-use function str_replace;
 use Symfony\Component\Finder\Finder;
+use function array_filter;
+use function array_keys;
+use function current;
+use function dirname;
+use function exec;
+use function extension_loaded;
+use function file_get_contents;
+use function implode;
+use function in_array;
+use function iterator_to_array;
+use function KevinGH\Box\FileSystem\canonicalize;
+use function KevinGH\Box\FileSystem\chmod;
+use function KevinGH\Box\FileSystem\dump_file;
+use function KevinGH\Box\FileSystem\make_tmp_dir;
+use function KevinGH\Box\FileSystem\mkdir;
+use function realpath;
+use function sprintf;
+use function str_replace;
 use function trim;
+use const DIRECTORY_SEPARATOR;
+use const PHP_EOL;
+use const STDOUT;
 
 /**
  * @covers \KevinGH\Box\Box
