+++ tests/Console/Command/ConfigurableTest.php
@@ -16,17 +16,17 @@ namespace KevinGH\Box\Console\Command;
 
 use Assert\InvalidArgumentException;
 use Closure;
-use KevinGH\Box\Configuration;
-use function KevinGH\Box\FileSystem\dump_file;
-use function KevinGH\Box\FileSystem\touch;
+use KevinGH\Box\Configuration\Configuration;
+use KevinGH\Box\Configuration\NoConfigurationFound;
 use KevinGH\Box\Json\JsonValidationException;
-use KevinGH\Box\NoConfigurationFound;
 use KevinGH\Box\Test\CommandTestCase;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\NullOutput;
 use Symfony\Component\Console\Output\OutputInterface;
+use function KevinGH\Box\FileSystem\dump_file;
+use function KevinGH\Box\FileSystem\touch;
 
 /**
  * @covers \KevinGH\Box\Console\Command\ConfigurableCommand
@@ -62,7 +62,7 @@ class ConfigurableTest extends CommandTestCase
 
         $output = new NullOutput();
 
-        /** @var Configuration $config */
+        /** @var \KevinGH\Box\Configuration\Configuration $config */
         $config = Closure::bind(
             function (ConfigurableCommand $command, InputInterface $input, OutputInterface $output): Configuration {
                 return $command->getConfig($input, $output);
@@ -152,7 +152,7 @@ class ConfigurableTest extends CommandTestCase
 
         $output = new NullOutput();
 
-        /** @var Configuration $config */
+        /** @var \KevinGH\Box\Configuration\Configuration $config */
         $config = Closure::bind(
             function (ConfigurableCommand $command, InputInterface $input, OutputInterface $output): Configuration {
                 return $command->getConfig($input, $output, true);
